# docs/openapi.yaml
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
...
  /decks:
    post:
      tags:
      - decks
      summary: Create Deck
      operationId: create_deck_decks_post
      parameters:
      - name: access_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - name: refresh_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Refresh Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeckCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - decks
      summary: Get Decks
      operationId: get_decks_decks_get
      parameters:
      - name: parent_id
        in: query
        required: false
        schema:
          type: string
          format: uuid
          title: Parent Id
      - name: exclude_paused
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Exclude Paused
      - name: exclude_archived
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Exclude Archived
      - name: access_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - name: refresh_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Refresh Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeckOut'
                title: Response Get Decks Decks Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /decks/tree:
    get:
      tags:
      - decks
      summary: Get Decks Tree
      operationId: get_decks_tree_decks_tree_get
      parameters:
      - name: access_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - name: refresh_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Refresh Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckTree'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /decks/{deck_id}:
    get:
      tags:
      - decks
      summary: Get Deck
      operationId: get_deck_decks__deck_id__get
      parameters:
      - name: deck_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Deck Id
      - name: access_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - name: refresh_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Refresh Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - decks
      summary: Update Deck
      operationId: update_deck_decks__deck_id__patch
      parameters:
      - name: deck_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Deck Id
      - name: access_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - name: refresh_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Refresh Token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeckUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeckOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - decks
      summary: Delete Deck
      operationId: delete_deck_decks__deck_id__delete
      parameters:
      - name: deck_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Deck Id
      - name: access_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - name: refresh_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Refresh Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /decks/import:
    post:
      tags:
      - decks
      summary: Import Deck
      operationId: import_deck_decks_import_post
      parameters:
      - name: format
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/ImportFormat'
          default: repeater
      - name: access_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - name: refresh_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Refresh Token
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_import_deck_decks_import_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /decks/{deck_id}/export:
    get:
      tags:
      - decks
      summary: Export Deck
      operationId: export_deck_decks__deck_id__export_get
      parameters:
      - name: deck_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Deck Id
      - name: include_sub_decks
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Include Sub Decks
      - name: access_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Token
      - name: refresh_token
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Refresh Token
      responses:
        '200':
          description: A downloadable JSON file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    Body_import_deck_decks_import_post:
      properties:
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
      - file
      title: Body_import_deck_decks_import_post
...
    DeckCreate:
      properties:
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          default: ''
        parent_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Parent Id
      type: object
      required:
      - name
      title: DeckCreate
    DeckNode:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        children:
          items:
            $ref: '#/components/schemas/DeckNode'
          type: array
          title: Children
        card_count:
          type: integer
          title: Card Count
        depth:
          type: integer
          title: Depth
        is_paused:
          type: boolean
          title: Is Paused
      type: object
      required:
      - id
      - name
      - children
      - card_count
      - depth
      - is_paused
      title: DeckNode
    DeckOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        parent_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Parent Id
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        is_paused:
          type: boolean
          title: Is Paused
        is_archived:
          type: boolean
          title: Is Archived
        is_root:
          type: boolean
          title: Is Root
        path:
          items:
            type: string
          type: array
          title: Path
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - user_id
      - name
      - is_paused
      - is_archived
      - is_root
      - path
      - created_at
      - updated_at
      title: DeckOut
...
    DeckTree:
      properties:
        decks:
          items:
            $ref: '#/components/schemas/DeckNode'
          type: array
          title: Decks
        total_decks:
          type: integer
          title: Total Decks
        tree_depth:
          type: integer
          title: Tree Depth
      type: object
      required:
      - decks
      - total_decks
      - tree_depth
      title: DeckTree
    DeckUpdate:
      properties:
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
          default: ''
        parent_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Parent Id
        is_paused:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Paused
        is_archived:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Archived
      type: object
      title: DeckUpdate
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ImportFormat:
      type: string
      enum:
      - repeater
      - mochi_markdown
      title: ImportFormat
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError